Modify any helper for multiple args (or condition)
While working with templates, it's often helpful to be able to output
content if any one of several arguments is valid. (For instance, only
show a product tab if there is warranty info OR shipping info. Only
show social icons wrapper if any one of the icons is present, etc.)

- Extend the any helper so it will work varying arity
- Set up some tests to work against the any document